backend/requirements.txt
# Core Django
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-allauth==0.57.0
argon2-cffi==23.1.0
django-ratelimit==4.1.0
django-simple-captcha==0.6.0

# Celery & Redis
celery==5.3.4
redis==5.0.1
django-celery-beat==2.6.0
django-celery-results==2.5.1

# Image handling
Pillow==10.1.0

# Environment variables
python-decouple==3.8
python-dotenv==1.0.0

# API documentation
drf-spectacular==0.27.0

# Admin UI
django-jazzmin==2.6.0

# Utilities
qrcode==7.4.2
requests==2.31.0
python-dateutil==2.8.2

# Crypto (for wallet address validation)
web3==6.11.3
tronpy==0.4.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring & Logging
sentry-sdk==1.39.1
